pyproject.toml
[project]
name = "bzero-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "celery[redis]>=5.6.0",
    "fastapi>=0.121.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.12",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-socketio[asyncio]>=5.11.0",
    "redis>=5.2.1",
    "sqlalchemy[postgresql-asyncpg]>=2.0.44",
    "uuid-utils>=0.11.1",
    "uvicorn[standard]>=0.38.0",
]

[project.scripts]
dev = "bzero.main:dev"

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.14.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bzero"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/bzero",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiohttp>=3.13.2",
]
